Split up protocol subscription controller
