+++ app/controllers/api/v1/cookie_and_jwt_api/cookie_and_jwt_api_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module CookieAndJwtApi
+      class CookieAndJwtApiController < ApiController
+        include ::Concerns::IsLoggedInWithAnyMethod
+      end
+    end
+  end
+end
