+++ config/routes.rb
@@ -52,10 +52,14 @@ Rails.application.routes.draw do
 
       # JWT and Cookie apis
       # TODO in V2 API, rename to people.
-      resources :person do
-        collection do
-          get :me
-          put :update
+      scope module: :cookie_and_jwt_api do
+        resources :protocol_subscriptions, only: [:create, :show]
+
+        resources :person, only: [] do
+          collection do
+            get :me
+            put :update
+          end
         end
       end
 
@@ -69,7 +73,7 @@ Rails.application.routes.draw do
         end
         resources :auth_user, only: [:create]
 
-        resources :protocol_subscriptions, only: [:create, :show] do
+        resources :protocol_subscriptions, only: [] do
           collection do
             get :mine
           end
@@ -87,8 +91,6 @@ Rails.application.routes.draw do
         resources :team, only: [:show], param: :group
         resources :organization, only: [:show], param: :group
       end
-
-
 
       # Public APIs
       resources :statistics, only: [:index]
